+++ blobUtils.hpp
@@ -3,7 +3,7 @@
 #include <opencv2/core/core.hpp>
 #include "cvblob.h"
 
-float diff(cv::Point &a, cv::Point &b);
+double diff(cv::Point &a, cv::Point &b);
 void blobToPointVector(cvb::CvBlob &p,std::vector<cv::Point> &points);
 void createLarvaROI(cv::Mat &frame, cv::Mat &ROI, cvb::CvBlob &blob);
 
@@ -15,11 +15,11 @@ void createLarvaContourPoints(cv::Mat &lrvROI,
                               cvb::CvBlob &blob,
                               int type=CV_8UC1);
 
-float getGreyValue(cv::Mat &larvaROI, cvb::CvBlob &blob,cv::Mat &grey_frame);
+double getGreyValue(cv::Mat &larvaROI, cvb::CvBlob &blob,cv::Mat &grey_frame);
 
-float getPerimeter(cvb::CvBlob &blob);
+double getPerimeter(cvb::CvBlob &blob);
 
-float getSurroundingSize(cv::Point &point, cvb::CvBlob &blob,cv::Mat &grey_frame);
-float plotAngle(cvb::CvBlob *blob,cv::Mat &ROIimg,int PAD=0);
-float angle( cv::Point &pt1, cv::Point &pt0, cv::Point &pt2 );
+double getSurroundingSize(cv::Point &point, cvb::CvBlob &blob,cv::Mat &grey_frame);
+double plotAngle(cvb::CvBlob *blob,cv::Mat &ROIimg,int PAD=0);
+double angle( cv::Point &pt1, cv::Point &pt0, cv::Point &pt2 );
 //#endif
